PR changes renaming and restructuring
